Allow op_planner to return a close enough path
Allow the op_planner to find a close enough waypoint for the goal.
The aim is to create a "two objective" target, the priority is to find
a good waypoint to the goal, if it can't be satisfied, then it
fall-back to the second objective that is finding the closest waypoint
below a threshold.

Change-Id: I0f4c09a6a86f23d93271bc8e5e9bcb74c526550e
Issue-Id: SCM-1096
Signed-off-by: Luca Fancellu <luca.fancellu@arm.com>
Co-authored-by: Senthil Ramakrishnan <senthil.ramakrishnan@arm.com>